Removing points from entities
Previously points were being stored inside entities, now there is
only the set of points of st->ot. Entities are ids to objects
that can be ennumerated over this point set.

Also a new interface for subset iterator was added (partially) that does
not require a subset_it_t struct.