Adding better interface for search results
It has been a very common use case to have a backtracking algorithm
implemented, and want the results in different formats or different
destinations. Usual cases are:

   - Print to stdout what the algorithm is finding.
   - Store an array of equally sized sequences, in memory and in a
     file.
   - Store the full search tree in memory. Cache it on a file either
     as a tree or as a variably sized sequence array.

The problem is these are not mutualy exclusive for a single algorithm,
meaning we can't hardcode a specific use case in the algorithm's
implementation.

For instance we might want an array of integers for thrackles, but we
may want to see them as a tree too.  Maybe we want to create a file but
at the same time store it in memory to avoid disk I/O. A file might be
better to seek a result in point_set_mode, but having the full array in
RAM may be better for sequential searches over all results.

All this is what motivated the creation of a sequence_store_t type, with
a unified API. The idea is to allow choosing the destination from the
algorithm's callsite. Meanwhile, inside the backtracking algorithm a few
functions that work for any destination or format are called. Even
though we may incurr in some computational costs for allowing this
flexibility, we intend to be explicit about these overheads.

This allows people who implement the algorithm to state the intended and
optimal use in its documentation, but not force the user to modify the
internals of it whenever they need another use case. Which can clutter
and make the algorithm harder to understand.

This is the first commit towards the goal of this unified API. What has
already being implemented is:

   - Unified API for collections of integer sequences of fixed length.
   - Memory pool, file or stdout as destination for such sequences.
     *See generate_egde_disjoint_triangle_sets() and
      iterate_thrackles_backtracking ().

   - Separate API for tree creation in memory with a simple per node
     interface inside the algorithm. It still uses 2 functions, we want 1
     in the end.
     *See matching_decompositions_over_complete_bipartite_graphs()

   - NOTHING HAS BEN THROUGHLY TESTED, it's likely some fix commits will
     come next. This is more like a start for the API design.