Fix performance of get_single_thrackle()
The algorithm was not actually the one we had before because even though
we were not expanding nodes that produced sequences not in ascending
order, we were checking them when updating S_l and calling expensive
functions on them like have_intersecting_segments().

Also used a continue; to reduce a little bit the output instruction
count.