TEST: Compare number of nodes to reach first thrackle
This code has a test on search.c to count how many nodes are explored in
the search tree before reaching the first thrackle

A new function single_thrackle_slow() was added, it's like
single_thracle() but does not enforce the constraint that the resulting
thrackle must have an ascending order. This is done for comparison
purposes.