Add a callback to sequence_store_t's API
The API of sequence_store_t was designed for backtracking algorithms
that explore some search tree. It's common to want to iterate the tree
of the same algorithm but perform different actions (i.e. print
sequences in different formats, count stuff, generate new structures
based on the tree etc.), for each of these actions we want to use the
same backtracking code.

If the search tree is small and fits in RAM (or a file) we can run the
search, output the tree, and then iterate it as needed to do what we
want. The callback API added here, is useful when this is not the case
and the search tree is too big.

One example is the search tree for the current version of
matching_decompositions_over_K_n_n() that would require 839 GiB.