Tried a lot of things to improve point auto positioning
Most of the attempts were unsuccessful but there is code on how I tried
them that may be useful for historical purposes. I'm commiting this
before starting a cleanup and discussion of what I tried.

Also we had several bugs before. Sorting around a point didn't work on
all cases, and we were adding forces to the incorrect index at the
beginning of the sweep across pairs of segments. These fixes made the
angular force version symmetric too.