Use a worker thread to arrange the points
Currently we only have one worker thread, a queue would be better
because now when the user moves before the algorithm finishes, we have
to wait on it, to start it again. Also, threads seem to take a long time
to cancel for n>8 is this because of the O(n^2) step time? or are we
doing too many frees to end a thread?